Fix the demo crash when it fail to execute before init_workers.
The root reason is that the wdata is initialized in init_workers and there is no check of wdata in exit_cleanup.
So the program will crash if it fail to execute before init_workers.

+++ demo/server.c
@@ -1195,9 +1195,11 @@ void stop_processes(void)
 {
 	int i;
 
-	for(i = 0; i < num_workers; i++) {
-		if (wdata[i].process)
-			kill(wdata[i].process, SIGTERM);
+	if (wdata) {
+		for(i = 0; i < num_workers; i++) {
+			if (wdata[i].process)
+				kill(wdata[i].process, SIGTERM);
+		}
 	}
 }
